+++ ops_lt/src/m_ops_src_char.f90
@@ -40,9 +40,9 @@ PARAMETER    (ROUTINENAAM = 'ops_src_char')
 
 ! SUBROUTINE ARGUMENTS - INPUT
 LOGICAL,   INTENT(IN)                            :: f_z0user                   ! user overwrites z0 values from meteo input
-REAL*4,    INTENT(IN)                            :: z0_user                    ! roughness length specified by the user [m]
-INTEGER*4, INTENT(IN)                            :: xb                         ! x-coordinaat van huidige bron in buffer
-INTEGER*4, INTENT(IN)                            :: yb                         ! y-coordinaat van huidige bron in buffer
+REAL,      INTENT(IN)                            :: z0_user                    ! roughness length specified by the user [m]
+INTEGER,   INTENT(IN)                            :: xb                         ! x-coordinaat van huidige bron in buffer
+INTEGER,   INTENT(IN)                            :: yb                         ! y-coordinaat van huidige bron in buffer
 TYPE (TApsGridInt), INTENT(IN)                   :: z0nlgrid                   ! map of roughness lengths in NL [m]
 TYPE (TApsGridInt), INTENT(IN)                   :: z0eurgrid                  ! map of roughness lengths in Europe [m]
 
@@ -50,7 +50,7 @@ TYPE (TApsGridInt), INTENT(IN)                   :: z0eurgrid                  !
 TYPE (TError), INTENT(INOUT)                     :: error                      ! error handling record
 
 ! SUBROUTINE ARGUMENTS - OUTPUT
-REAL*4,    INTENT(OUT)                           :: z0_src                     ! roughness length at source; from z0-map [m]
+REAL,      INTENT(OUT)                           :: z0_src                     ! roughness length at source; from z0-map [m]
 
 !-------------------------------------------------------------------------------------------------------------------------------
 !
