+++ ops_lt/src/m_ops_write_progress.f90
@@ -36,16 +36,16 @@ CHARACTER*512                                    :: ROUTINENAAM                !
 PARAMETER    (ROUTINENAAM = 'ops_write_progress')
 
 ! SUBROUTINE ARGUMENTS - INPUT
-REAL*4,    INTENT(IN)                            :: progress                   ! percentage of progress reached
+REAL,      INTENT(IN)                            :: progress                   ! percentage of progress reached
 CHARACTER*(*), INTENT(IN)                        :: formatstring               ! formatstring for writing progress
-INTEGER*4, INTENT(IN)                            :: numbs                      ! number of characters which have to be backspaced
+INTEGER,   INTENT(IN)                            :: numbs                      ! number of characters which have to be backspaced
                                                                                ! in order to remain at the same position of the screen 
 
 ! SUBROUTINE ARGUMENTS - I/O
-INTEGER*4, INTENT(INOUT)                         :: memdone                     
+INTEGER,   INTENT(INOUT)                         :: memdone                     
 
 ! LOCAL VARIABLES
-INTEGER*4                                        :: idx                        ! do-loop index
+INTEGER                                          :: idx                        ! do-loop index
 
 !-------------------------------------------------------------------------------------------------------------------------------
 !
